move abstract Navigator class, remove circular...
imports between ELONavigator and base class. Fiddle with relitive import
paths.
