+++ packages/db/src/core/types/contentNavigationStrategy.ts
@@ -6,7 +6,7 @@ import {
   StudySessionNewItem,
   StudySessionReviewItem,
   UserDBInterface,
-} from '..';
+} from '../index';
 import { Navigators } from '../navigators';
 import ELONavigator from '../navigators/elo';
 
@@ -28,33 +28,4 @@ export interface ContentNavigationStrategyData extends SkuilderCourseData {
    by the implementing class's constructor at runtime.
   */
   serializedData: string;
-}
-
-/**
- * A content-navigator provides runtime steering of study sessions.
- */
-export abstract class ContentNavigator implements StudyContentSource {
-  /**
-   *
-   * @param user
-   * @param strategyData
-   * @returns the runtime object used to steer a study session.
-   */
-  static create(
-    user: UserDBInterface,
-    course: CourseDBInterface,
-    strategyData: ContentNavigationStrategyData
-  ): ContentNavigator {
-    const implementingClass = strategyData.implementingClass;
-
-    switch (implementingClass) {
-      case Navigators.ELO:
-        return new ELONavigator(user, course);
-      default:
-        throw new Error(`Unknown implementing class: ${implementingClass}`);
-    }
-  }
-
-  abstract getPendingReviews(): Promise<(StudySessionReviewItem & ScheduledCard)[]>;
-  abstract getNewCards(n?: number): Promise<StudySessionNewItem[]>;
 }
